+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Set up expo router [POSTS-005] #5.
 - Add person code validation tests [POSTS-004] #4.
 - Created CHANGELOG [POSTS-003] #3.
 - Added login screen with persons code validation [POSTS-002] #2.